[tools][fuzz] Implement v2 fuzzer workflows
For v2 (i.e. CFF) fuzzers, the arguments passed to an undercoat
`run_fuzzer` call now map to a series of appropriate calls to `ffx
fuzz`, along with various other logic to transparently touch up the
output of libFuzzer and move files to/from expected locations.

Along the way, the Fuzzer unit tests have been significantly expanded to
be as close as possible to the coverage of the e2e tests.

Change-Id: I8e8d8ac10992e8621a86097ad3837437ce21c362
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718407
Reviewed-by: Aaron Green <aarongreen@google.com>
Commit-Queue: Cameron Finucane <eep@google.com>