[recovery] Re-enable boot simulator test
This was disabled for the CFv2 migration. It depends on serial log
output which was implicitly emitted in CFv1. An extra change was
required to re-enable serial logging for CFv2.

This change also moves the key log line to the setup method of
AppAssistant, which is further along in app initialization than the
previous "recovery:started" message.

Bug: b/241963587
Test: fx test -v recovery_simulator_boot_test
Change-Id: I0a98d93a30207a66c234b3e2d95e869b6e7ee108
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719028
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>