[media/vnext] delete vnext prototype code and FIDL
This CL deletes the media/vnext prototype code and associated FIDL
definitions that are not otherwise in use. Given that the API design
is changing and the fact that this code uses HLCPP, this prototype
will not be useful in this form. Some of this code will reappear in
a different form as V2 work proceeds.

Test: deletion only, no new behavior
Change-Id: I028f74523c6510a1b83adf55a2933b9ea5cb999b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723465
Reviewed-by: Tom Bergan <tombergan@google.com>
API-Review: Dale Sather <dalesat@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>