[dart] Turn off Dart documentation generation.
This is currently blocking sandboxing for infrastructure because it
calls out to the internet (fxb/108151). We need to figure out whether the documentation should be kept around or removed, but for now I'm turning off updates to unblock infra.

Other reasons why I think this can be turned off safely:

  - Very low usage according to Google Analytics https://screenshot.googleplex.com/4MQFTjyE29iLkQe.png
  - We are eventually moving Dart support out of fuchsia.git https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0176_disallow_new_dart_programs?hl=en
  - The generated documentation is not particularly high quality.

Change-Id: I8b41d26bed0f170e15435d571d8e536aa78243d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719250
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Alexander Biggs <akbiggs@google.com>
Reviewed-by: Kendal Harland <kjharland@google.com>