[netdevice-migration] Fix race in releasing netbufs
We are signaling to netdevice-core that a transaction is
completed and that a buffer is available before actually releasing
that buffer. This can result in netdevice-core sending a new
buffer before the buffer is returned to pool, resulting in
a false contract violation.

Resolve by signaling to netdevice-core only after releasing the buffer
back to the pool.

Change-Id: I8b5545e42c3643b7d565c270d31223be869b54af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722967
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Joon-Hyuk Han <hanjh@google.com>