[pm] Improve error message when ABI revision is not specified
We now need the ABI revision when building a package (either passed in
through the CLI, or manually created in the package directory). This
improves the error to make it easier to users to understand how that
this is required.

Change-Id: I48b311bdd8e1784d024d950b83d570b063e528b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724008
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Buckland <jbuckland@google.com>