[icu][build] Adds third_party_icu_headers template
This is used for the ability to place the ICU library
into a directory other than the default //third_party/icu.

Some products need to use the ICU library from a different
directory, and this template makes the changes necessary
for that to be possible.

Bug: 108544
Change-Id: I6ff817703034a1b47d34b5261683e5eab7a813eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720284
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>