[storage] change DELETED watch event name to "."
Change the name returned when the DELETED watch event is sent to "."
instead of the name of the directory.

For all other watch events, the name sent by the event doubles as the
relative path from the directory on which the watcher was set. This
brings the DELETED watch event in line with that behavior to make it
more consistent.

This cl also adds support for the DELETED watch event to fatfs and
minfs.

Fixed: 76762
Test: fx test watcher-tests
Change-Id: I0dbd8bb0d4ab13655aae701fbc937e2eb26fbf85
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722051
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
API-Review: Chris Suter <csuter@google.com>