[pci] Migrate PCI from Banjo to FIDL under the hood
This changes the internal implementation of the ddk::Pci class to use
FIDL instead of Banjo. It also changes all users of the class to bind to
the FIDL PCI nodes in the device tree, and updates some instances of PCI
topological paths in the source code.

There are a few remaining PCI Banjo users that are not migrated by this
CL because they do not use ddk::Pci, notably iwlwifi, e1000, and
brcmfmac.

Test: booted Chromebook, NUC, and emulators
Bug: 99914
Change-Id: I9885986ea3f8b0b54052319d339398c73625bab2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714013
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: Ian Fisher <iafisher@google.com>