[fidl-fuchsia-pkg-ext] Add serve_fidl_iterator_from_stream
Serves a FIDL iterator from a Stream of payloads. Useful for when the
payloads are generated concurrently with the serving of the iterator
(e.g. the fuchsia.pkg/NeededBlobs.GetMissingBlobs iterator will
be served concurrently with the handling of
fuchsia.pkg/NeededBlobs.OpenBlob when subpackages are implemented).

Bug: 109245
Change-Id: I454be8650e65c10f470f90c8c4cdaa790f576a22
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723537
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>