[testrunner] Add QEMU and home directory mounts
Adds a couple more mounts to keep existing host tests happy. The usage
of the current user's home directory is particularly hacky - we maintain
isolation by replacing it with a tmpfs in the sandbox, but the usage of
the /etc/passwd to find the name of this directory by a couple of host
tests is not ideal.

Bug: b/211504655
Change-Id: Ic010a76247be678c5b4b88375f66d178f8f83fb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723422
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>