[debugger] Use CommandContext in nouns.
Uses the new CommandContext object in the noun processing. This creates
a new ConsoleCommandContext just before calling into the noun system to
limit the scope of this patch and keep the rest of the system running.

The rest of the command processing will be converted in subsequent
passes. Then the CommandContext will be passed in rather than created
for the noun system by DispatchCommand.

Change-Id: I34d4d7da7f6580a4cf68660b1dcf51cb611c40de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723479
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>