[zither] Fix name utility inconsistency
zither.(Upper|Lower)CaseWithUnderscores() previously included the
library name as tokens, something zither.UpperCamelCase() does not. The
inclusion of the library name is rather C-specific (for want of a proper
library namespace), so this change removes the library name support in
the zither package and creates a wrapper with this support in the C one.

Bug: 51002
Change-Id: I228c88464d30a8e9715d4afc9e803b7406efbfd9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723782
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Brett Wilson <brettw@google.com>