[kernel] Log more when a process critical to root job dies
Any crash in a process that will bring the system down when it dies
is an important thing to log about in as much detail as possible.

This adds direct kernel logging about every exit from a critical
process, so that in all paths at least the retcode will be logged.
In the exception path, it adds full logging of exception and
register details for any exception fatal to a critical process.

Bug: 108859
Change-Id: I967c614739039ddb270820be5abd5bd505e76d0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723373
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>