[fshost] implement basic minfs/zxcrypt support
Add support for unsealing/formatting zxcrypt and mounting minfs on top
of that. For now we always use null keys for zxcrypt, which is fine for
the test, since that also hard-codes null keys, but the rewrite needs to
support getting the hardware keys as well. fxbug.dev/109067 tracks the
support we need for that.

Bug: 87511
Bug: 109067
Test: fx test fshost_integration_tests
Test: fx test fshost-rust-test
Change-Id: I99ce7ec19ed54b2d50c8f7b6da420b41bbb5a62f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723369
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>