[kernel][vm] Specify is_zero_range on writeback begin
ZX_PAGER_OP_WRITEBACK_BEGIN supports optionally passing in
ZX_VMO_DIRTY_RANGE_IS_ZERO through the data argument. This indicates
that the user pager wants to write back the specified range as zeroes,
and so any new non-zero content (actual pages) created in this range
must be preserved after the writeback. In other words, any committed
dirty pages in the specified range must be left dirty.

This option is meant to be used when the user pager is writing back a
zero range that it queried with zx_pager_query_dirty_ranges(). A page
that is dirtied after the query but before the writeback is started
should remain dirty. With an awaiting clean zero range now allowing
committed pages in the range, it was possible for committed pages to be
prematurely marked clean and data to be lost when they were evicted.

Test: core-pager-writeback
Bug: 108785
Bug: 63989
Run-All-Tests: true
Change-Id: If23d2e7a390057ac1cddf967307c6cf0159381ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723375
Reviewed-by: Adrian Danis <adanis@google.com>
Fuchsia-Auto-Submit: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>