[fidl][example-tester] Create example-tester lib
The example-tester is a lightweight testing framework for testing FIDL
canonical examples. The key usefulness of the framework is that test
assertions may be performed on log output only. This is useful because
the primary purpose of canonical examples is not to make for good,
complete, robust tests, but rather to serve as terse and readable code
samples. In this light, a test framework that encourages logging of
important code checkpoints, but does not produce any other output that
would clutter up example files, is desirable

Test: fx test example_tester -r --count 100
Bug: 107613
Change-Id: Ib1a713ae23facfd0dfb4988620b5ee6cf169daea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/694933
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Matt Hamrick <mjhamrick@google.com>