[pkg-cache] Update tests per RFC6761
RFC6761 defines special-use domain names, including domains to be used
in documentation. `pkg-cache` unit tests encode fuchsia.com, which is a
name that may exist in DNS. It can be unclear whether reliance on the
function of that domain within DNS is relevant to the test.

Base packages are always served from fuchsia.com, this change makes it
possible for the test to use a different hostname.

Fixed: 108319
Change-Id: I8db6866668dea550423fea8809ec1fa75f81fed7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723464
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>