[ffx][daemon] Graceful exit always waits a bit
When adding the mechanism where the daemon gracefully shuts down when
the socket file is deleted, we accidentally made it so that this
bypasses the wait from a graceful shutdown message intended to allow
existing messages to finish processing (including and especially the
reply to the quit() message), because it would delete the file
immediately and then the delete handler wouldn't wait.

This moves the delay logic into the main loop itself, making it so that
*all* paths through the quit logic go through the same process,
including the brief wait.

Fixed: 109242
Change-Id: I1599457ef67cb109fb4fd00849a2ad2bc83190f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724084
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>