[audio][services][mixer] Generalize PacketQueueProducerStage
This CL renames PacketQueueProducerStage to ProducerStage, then
refactors that class so its source stage can be either of these classes:

- SimplePacketQueueProducerStage
- SimpleRingBufferProducerStage

These classes no longer inherit directly from ProducerStage.

As part of this refactor, PacketQueueProducerStage::Command was split
into two types, which are now sent over two separate queues:

- ProducerStage::Command (includes Start and Stop)
- SimplePacketQueueProducerStage::Command (includes PushPacket and
  Clear)

Most of this change is mechanical fallout from this refactor. The core
functional change is in ProducerStage, which is now simpler because it
not longer needs to deal with PushPacket and Clear.

Bug: 87651
Change-Id: I45df628a9d98cae80fba7a17544350957332c95e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723310
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>