[settings] Add custom watch function for audio
Because we increment the modification counters stored in AudioInfo for
every change, clients are notified for every change, even if it is
duplicated. We need to keep the modification counters for earcons to
properly play on voice change, so a change function is instead added to
the Watch fidl handler to only respect stream changes.

Test: fx test setui_service_tests, audio_integration_tests
Multiply: audio_integration_tests
Fixed: 107775

Change-Id: I1f0706e38ad96a999c86e135dcb8913a2e8352d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722022
Reviewed-by: William Xiao <wxyz@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Eric Rahm <erahm@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>