[Settings] Reland "Convert Display to use jobs"
This relands fxr/707813 with a fix for a loop where a client that
watched the light sensor with a delta of 0 would continually receive
updates on their hanging get. See the bug for details on the investigation.

The diff between the original CL and this fix can be found at fxr/722965.

Converts the `Display` settings to use jobs. Also removes an unused
hanging get function.

Fixed: 79041
Test: fx test setui_service_tests
      manually verified through device logs that there's no loop of
      light sensor watches
Change-Id: I5e5711e7888657695db6516c46841039dbb5619e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722963
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: Eric Rahm <erahm@google.com>
Commit-Queue: William Xiao <wxyz@google.com>