[fidl][cpp] rename Unknown(Interaction->Method).
Rename FIDL Status::UnknownInteraction to Status::UnknownMethod.

This status is only used in clients to identify when a method call made
by the client was unknown to the server. Unknown events and unknown
methods on the server are handled using a handler interface rather than
an error status.

This renaming is intended to make naming more consistent across
different parts of unknown interaction handling. The handler interfaces
are named UnknownMethodHandler and UnknownEventHandler, so naming the
corresponding error UnknownMethod is more consistent.

This is part of a set of changes which will make names more consistent
across all languages.

Bug: 88366
Change-Id: Ic31c668cbb0a742a88591870ac8bd93220bb41e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722632
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>