[build] Have assembled_system() emit image metadata
Currently, all relevant image metadata is specified outside of
assembled_system(). This is largely due to the fact that image metadata
is quite a mess and contains a lot of information that is out-of-band to
the template. This change moves the canonical image metadata into the
template, keeping the wonkiness in the user-defined domain with a new
`image_metadata_overrides` parameter.

Bug: 81569
Change-Id: I9a0d30401213d6d62c137dc6036e7f1543e592b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723490
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>