[kazoo] Replace golden file testing scheme
The previous way concatenated every kazoo output into a single, very
large text file and mediated golden diffing through that with a bespoke
scheme. The sheer size of this file and its presentation as raw text
made it rather difficult to grok what the kazoo outputs actually were -
and updates had to be manual.

In this change we remedy that by replacing that infrastructure with now
canonical `golden_files()`, which allows us to separate out the files
(presenting them as proper source code), remove the bespoke logic, and
update them via a simple build or more directly with
`fx check-goldens`*. We check in the goldens under
//zircon/vdso/goldens. (Without further reorganization, this
is the most convenient place to put them as a target in
//zircon/vdso/BUILD.gn is what currently generates most of the files in
question.)

*In this case the minimal way to update them is `fx check-goldens vdso`.

Change-Id: I3e176074f05e838786c9e8454801a3671d5e4225
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724314
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>