[scripts][sdk]: Fix depfile generation in merger script.
Fix the depfile generated by the SDK merger script.
Before this CL, the script would include paths to temporary
files in the output, which polluted the Ninja deplog for no
good reason.

This also fixes the --hermetic-inputs-file behavior, by
not touching the output directory, or generating an archive,
when this flag is used. Its content is also fixed since it
had the same issue as the depfile one.

+ Refactor the script in several ways to prepare for a
  future CL that will add support for merging multiple
  inputs at the same time. I.e.

  - Introduce MergeState class to hold global state
    that is shared for all merge operations (currently
    only one), and ensure that temporary input and
    output files are _not_ listed in the depfile.

  - Introduce InputSdk and OutputSdk to model inputs
    and outputs for merge operations.

  - Introduce ElementMetadata class to model the
    content of an element's metadata file and perform
    operations on it (get type, get files, merge)

+ Use Python annotations to clarify the code and
  type-check it properly.

Bug: None

Change-Id: Iebe9e48a81be9047bba0a3da73b0df9dab3f9870
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722662
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>