[icu][build] Use third_party_icu_headers
This is used for the ability to re-target the use of
the ICU library, which is required for some products.

We need to replace the use of third_party/icu with
a variable root directory to be able to vary ICU use,
both in BUILD.gn files, and in C++ files where
the #include directives refer to header files by
their path.

Bug: 108544
Change-Id: I8733c05548105836244594c6eb3177643bdda706
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719663
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>