[system-updater] Update tests per RFC6761
RFC6761 defines special-use domain names, including domains to be used
in documentation. 'system-updater' unit tests encode fuchsia.com and
example.com, which are names that may exist in DNS. It can be unclear
whether reliance on the function of that domain within DNS is relevant
to the test.

By complying with the special-use domain names in these tests, there is
less cause for DNS concerns in this component.

Test: system-updater

Fixed: 109339
Change-Id: Id298af2b1aacddf91afd377e06dba5b9fd7258cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724619
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>