[kernel][sched] Remove `period` from deadline profiles.
Remove the `period` field from the kernel implementation of deadline
profiles.  Having a period separate from the profile's relative
deadline was a feature which was not really being used, and it was
making the generalization of deadline profile inheritance difficult.

So for now, remove the concept of deadline period from the kernel's
implementation of deadline profiles.  After this, we can either remove
the period from the profile definition entirely, or we can
re-introduce it once we figure out how to properly incorporate it into
deadline inheritance.

Change-Id: Ib6a2204da8c4556379dd1386d7bf6f55641fea24
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658372
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>