[docs] Add documentation of the IfThisThenThat checker
IfThisThenThat is a presubmit check that runs automatically. However, it
requires the user to understand that it exists as it operates on special
comments that need to be inserted into the code (most presubmit checks just operate
independently).

Because IFTTT requires some teaching, it needs to be documented. As
there is no current page addressing presubmit checks, this change adds
on and places the IFTTT instructions within.

NO_IFFT=This is example code, not actually IfThisThenThat directives.

Change-Id: I29f6ea04284d841275ad77ee2fee6a7588189294
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724597
Reviewed-by: Vivian Tran <viviantran@google.com>
Commit-Queue: Chris Lewis <cflewis@google.com>