[fidl][cpp] Remove deadline support in sync calls
If the deadline elapses, we do not close the channel and the response
from the server would not be recognized. Because deadlines are not used,
it is easier to remove them now while we discuss what's the best
strategy for sync deadlines.

Change-Id: Ifb32948c26bd65d1f8aa72d1069d896d2ccabf8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724825
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Zachary Stewart <zstewart@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>