[assembly_input_groups] New bundle
Introduce assembly input groups, which represent a collection
of artifacts that correspond to partial input to a system image.

A group is generated for each attribute supported by the input groups,
each of this groups represents the list of labels that are to
be used in the assembled_system() template.

An extra field, is the set of core realm shards, this are used
to generate a core realm that should be embedded in the base
package set.

Manually tested by locally replacing the usage in fuchsia/BUILD.gn
to point towards this label instead of doing lookup.

Test: fx build
Change-Id: Ia8ae45c39e76f2f6e11fc603937b2aa69e1afce9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708584
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>