[debug_data] Don't schedule processing task if processing is done
This removes one potential redundant write that could occur. This also
fixes a flake where writing to the summary could be attempted after the
directory is torn down.

The flake could occur depending on the order in which AddData,
DataFinished, TakeMapContents, and FinishProcessing methods are called.
AddData and FinishProcessing are called from thread 1, TakeMapContents
and DataFinished are called from thread 2.
In case the order is AddData -> TakeMapContents -> FinishProcessing ->
DataFinished, the dirty bit is still set when the processing task
finishes. This causes thread 2 to dispatch the data processing task
again. This could trigger attempting to write to the summary file again.
With the recent change to support isolated debug data directories, at
the time this write is attempted the directory may by in the process of
being torn down (as it is managed by a separate component).

Multiply: test_manager_test
Fixed: 107144

Change-Id: I51558fa196f0775e7a57910b6d54f9f314a13444
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/725369
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>