[icu][build] Use third_party_icu_headers
This is used for the ability to re-target the use of
the ICU library, which is required for some products.

We need to replace the use of third_party/icu with
a variable root directory to be able to vary ICU use,
both in BUILD.gn files, and in C++ files where
the #include directives refer to header files by
their path.

Bug: 108544
Change-Id: Iaf173093d35326540a66088d1c56b8475aec42d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720285
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>