[dhcpv6] Move logic out of assert_matches in tests
Move logic out of usages of `assert_matches` on `ClientState` in unit
tests. These `assert_matches` usages make test failures difficult to
debug because the "want" values are not printed.

The additional benefit of not keeping too much code inside
`assert_matches` is that we can rely on the auto-formatter to format
more code.

Change-Id: If59ba446ac8a70f59a64e2ee772c5c3263a25488
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724363
Commit-Queue: Tony Gong <gongt@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>