[dhcpv6] Add collected_advertise to all states
The collected Advertises need to be tracked also in Assigned and
Renewing states because it may be necessary to try another server
when Renewebing/Rebinding if no usable lease is found.

Also assert the value of `collected_advertise` in tests wherever it is
convenient to do so (which basically means whenever it is supposed to be
empty, or the want value is known due to being directly constructed by
the test or obtained by inspecting a previous `ClientState`).

Mechanical changes only, no change in behavior.

Change-Id: I308ff8c93013aeba9635ff3e4fafb93144795514
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723728
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Tony Gong <gongt@google.com>