[identity] Persist AccountMetadata in AccountManager
We were ignoring the account metadata which was sent (or not) as part of
provision_new_account api. This change enables AM to start persisting it
as part of the stored_account_list.
This will most likely get moved to AccountHandler in the future.

Fixed: 101664
Change-Id: Ide69f7a85a694a45bb9978b191c2ca1371a77bb4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724409
Commit-Queue: Angad Pal Singh <apsbhatia@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>