[audio] Add presentation delay methods to mixer Node
This is to first step to introduce the presentation delay contributions
for each node in a mix graph to compute and report the corresponding
latency and lead time requirements back to the mixer server and clients.

This change also includes a minor code cleanup to reorder the Node
methods for consistency, and rename src -> source.

Bug: 87651

Multiply: audio_services_mixer_fidl_unittests

Change-Id: I9bc2e671c2daed4222907422b781a2a4d71841c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/725169
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Alper Gungormusler <alperg@google.com>