[fidl][fidlc] Use one FIDL for error doc and test
Tests for specific errors are now able to have their test case FIDL
inputs factored out into separate files, thereby allowing that input to
be used both as a test case and as the includecode for the markdown
documentation of the error in question. This creates a single source of
truth for both error documentation and the tests it documents.

Test: fx test host_x64/fidlc-test
Bug: 108248
Change-Id: I107b6d092e955d223c3764fa21e662ba40cf999a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718043
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>