[fdf] Create DFv2-specific workstation_eng groups
Create the GN product `workstation_eng_dfv2`. This is identical to
`workstation_eng` except it doesn't include tests that are failing when
DFv2 is enabled. Also create the GN group
`bundles/buildbot:workstation_dfv2` which is identical to
`bundles/buildbot:workstation` but it does not include tests that are
failing when DFv2 is enabled. Also separate buildbot workstation tests
into two groups: one for tests that are included in host labels and one
for tests that do not.

Existing builder bots with DFv2 and workstation_eng will use these
counterparts in a future CL. This is in order to get builder bots that
are running DFv2 passing and monitored by the Fuchsia build gardeners in
order to prevent regressions in DFv2.

Bug: 105115
Change-Id: If78a7ea8fdf2a9d2c2fdb5dc354cfafb77a64171
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718666
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>