[cm] Migrate destruction tests to events_v2
This CL migrates the destruction tests to use events V2 per RFC-0121. The tests
are refactored to deal with limitations of subscribing to events from within
a nested component_manager by not expecting to receive events in a particular
order or a particular set of events. Instead, it tests for a subset because
we may receive more events than we're expecting because there is no way
to subscribe to only a certain set of events.

Run-All-Tests: true

Bug: 81980
Change-Id: Iae8547e9a61d918e96ac9a81a49a270c4175b45e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722524
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Gary Bressler <geb@google.com>