[compat] Wait for unbind to be replied to before calling release
Right now we call release under the assumption unbind completed right
away. In reality, unbind is my be compeleted asynchronously. This may
deadlock as we are blocking the dispatcher from running. We detect that
via a 10 second timeout. This is not ideal, but crashing in this case is
probably better than calling release incorrectly.

The proper fix for this situation will require the prepare_stop hook to
be implemented as described in go/tq-dfv2-dispatcher-shutdown. This will
allow us to trigger unbind durring prepare stop and release during stop.

Bug: 97457
Bug: 92196
Test: fx test compat_test
Change-Id: Ife2f841a24ef0c5d80eb4ec4f9b50dff19218fd8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724987
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>